requirements.txt
Django
djangorestframework
django-cors-headers
drf-api-logger
djangorestframework-simplejwt
djangorestframework-jwt
pillow
gunicorn
whitenoise
requests
xlrd